requirements.txt
streamlit==1.39.0
pandas==2.2.3
beautifulsoup4==4.12.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
matplotlib==3.8.3
seaborn==0.13.2
